added controllers/routes for reviews and notifications. Also updated update user controller function so profile picture can be updated. Still need to add input validation for reviews

+++ backend/controllers/userController.js
@@ -1,6 +1,8 @@
 const userModel = require("../model/userModel");
 const bcrypt = require("bcrypt");
 const venueModel = require("../model/venueModel");
+const notificationModel = require("../model/notificationModel");
+const reviewModel = require("../model/reviewModel");
 
 exports.login = (req, res, next) => {
     let email = req.body.email;
@@ -89,7 +91,7 @@ exports.deleteAccount = (req, res, next) =>{
     .then((venues) =>{
         if(venues){
             let venueIds = venues.filter(venue => venue.id)
-            Promise.all([venueModel.deleteMany({_id: venueId}), bookingModel.deleteMany({venueId: {$in: venueIds}}), userModel.findByIdAndDelete(userId)])
+            Promise.all([venueModel.deleteMany({_id: venueId}), bookingModel.deleteMany({venueId: {$in: venueIds}}), userModel.findByIdAndDelete(userId), notificationModel.deleteMany({for: userId}, reviewModel.deleteMany({reviewerId : userId}))])
             .then((deletedItems) => {
                 if(deletedItems){
                     if (!req.session) {
@@ -119,10 +121,11 @@ exports.deleteAccount = (req, res, next) =>{
 
 exports.updateUser = (req, res, next) =>{
     let password = req.body.password
+    let img = req.files.map(file => `/images/${file.originalname}`)
     bcrypt.hash(password, 10)
     .then((hashedPass) => {
     let userId = req.body.id
-    user.findByIdAndUpdate(userId, {password: hashedPass}, {runValidators: true})
+    userModel.findByIdAndUpdate(userId, {password: hashedPass, image: img}, {runValidators: true})
     .then((user) =>{
         if(user){
             res.status(200).json({success: "User updated successfully"})
